chore: rename folder sorting utility
