+++ src/gui/ChoiceBuilder/templateChoiceBuilder.ts
@@ -24,7 +24,7 @@ import {
 	placementSupportsEmbed,
 } from "../../types/linkPlacement";
 import { getAllFolderPathsInVault } from "../../utilityObsidian";
-import { sortFolderPathsByTree } from "../../utils/folderSorting";
+import { sortFolderPathsByTree } from "../../utils/folder-sorting";
 import { createValidatedInput } from "../components/validatedInput";
 import { ExclusiveSuggester } from "../suggesters/exclusiveSuggester";
 import { FormatSyntaxSuggester } from "../suggesters/formatSyntaxSuggester";
